requirements_api.txt
flask==2.3.0
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0
werkzeug==2.3.0
